Move reshape and flatten to NNVM
