--- src/operator/reshape.cu
@@ -1,23 +1,0 @@
-/*!
- * Copyright (c) 2015 by Contributors
- * \file flatten.cc
- * \brief
- * \author Bing Xu
-*/
-
-#include "./reshape-inl.h"
-
-
-namespace mxnet {
-namespace op {
-template<>
-Operator *CreateOp<gpu>(ReshapeParam param, int dtype) {
-  Operator *op = NULL;
-  MSHADOW_REAL_TYPE_SWITCH(dtype, DType, {
-    op = new ReshapeOp<gpu, DType>(param);
-  });
-  return op;
-}
-
-}  // namespace op
-}  // namespace mxnet
