+++ src/operator/tensor/matrix_op.cc
@@ -9,12 +9,84 @@
 
 namespace mxnet {
 namespace op {
+DMLC_REGISTER_PARAMETER(ReshapeParam);
 DMLC_REGISTER_PARAMETER(TransposeParam);
 DMLC_REGISTER_PARAMETER(ExpandDimParam);
 DMLC_REGISTER_PARAMETER(SimpleCropParam);
 DMLC_REGISTER_PARAMETER(SimpleCropAssignScalarParam);
 DMLC_REGISTER_PARAMETER(SliceParam);
 DMLC_REGISTER_PARAMETER(FlipParam);
+
+NNVM_REGISTER_OP(Reshape)
+.MXNET_DESCRIBE("Reshape input according to a target shape spec.\n"
+"The target shape is a tuple and can be a simple list of dimensions "
+"such as (12,3) or it can incorporate special codes that correspond "
+"to contextual operations that refer to the input dimensions.\n"
+"The special codes are all expressed as integers less than 1. These "
+"codes effectively refer to a machine that pops input dims off the "
+"beginning of the input dims list and pushes resulting output dims "
+"onto the end of the output dims list, which starts empty. The codes "
+"are:\n"
+"  0  Copy     Pop one input dim and push it onto the output dims\n"
+" -1  Infer    Push a dim that is inferred later from all other output dims\n"
+" -2  CopyAll  Pop all remaining input dims and push them onto output dims\n"
+" -3  Merge2   Pop two input dims, multiply them, and push result\n"
+" -4  Split2   Pop one input dim, and read two next target shape specs,\n"
+"              push them both onto output dims (either can be -1 and will\n"
+"              be inferred from the other\n"
+" The exact mathematical behavior of these codes is given in the "
+"description of the 'shape' parameter. All non-codes (positive "
+"integers) just pop a dim off the input dims (if any), throw it away, "
+"and then push the specified integer onto the output dims.\n"
+"Examples:\n"
+"Type     Input      Target            Output\n"
+"Copy     (2,3,4)    (4,0,2)           (4,3,2)\n"
+"Copy     (2,3,4)    (2,0,0)           (2,3,4)\n"
+"Infer    (2,3,4)    (6,1,-1)          (6,1,4)\n"
+"Infer    (2,3,4)    (3,-1,8)          (3,1,8)\n"
+"CopyAll  (9,8,7)    (-2)              (9,8,7)\n"
+"CopyAll  (9,8,7)    (9,-2)            (9,8,7)\n"
+"CopyAll  (9,8,7)    (-2,1,1)          (9,8,7,1,1)\n"
+"Merge2   (3,4)      (-3)              (12)\n"
+"Merge2   (3,4,5)    (-3,0)            (12,5)\n"
+"Merge2   (3,4,5)    (0,-3)            (3,20)\n"
+"Merge2   (3,4,5,6)  (-3,0,0)          (12,5,6)\n"
+"Merge2   (3,4,5,6)  (-3,-2)           (12,5,6)\n"
+"Split2   (12)       (-4,6,2)          (6,2)\n"
+"Split2   (12)       (-4,2,6)          (2,6)\n"
+"Split2   (12)       (-4,-1,6)         (2,6)\n"
+"Split2   (12,9)     (-4,2,6,0)        (2,6,9)\n"
+"Split2   (12,9,9,9) (-4,2,6,-2)       (2,6,9,9,9)\n"
+"Split2   (12,12)    (-4,2,-1,-4,-1,2) (2,6,6,2)\n")
+.set_num_inputs(1)
+.set_num_outputs(1)
+.set_attr_parser(ParamParser<ReshapeParam>)
+.set_attr<nnvm::FInferShape>("FInferShape", ReshapeShape)
+.set_attr<nnvm::FInferType>("FInferType", ElemwiseType<1, 1>)
+.set_attr<nnvm::FGradient>("FGradient", ElemwiseGradUseNone{"_backward_copy"})
+.set_attr<FCompute>("FCompute<cpu>", IdentityCompute<cpu>)
+.set_attr<nnvm::FInplaceOption>("FInplaceOption",
+  [](const NodeAttrs& attrs) {
+    return std::vector<std::pair<int, int> >{{0, 0}};
+})
+.add_argument("data", "NDArray", "Input data to reshape.")
+.add_arguments(ReshapeParam::__FIELDS__());
+
+
+NNVM_REGISTER_OP(Flatten)
+.describe(R"(Flatten input into 2D by collapsing all the higher dimensions.
+A (d1, d2, ..., dK) tensor is flatten to (d1, d2* ... *dK) matrix.)")
+.set_num_inputs(1)
+.set_num_outputs(1)
+.set_attr<nnvm::FInferShape>("FInferShape", FlattenShape)
+.set_attr<nnvm::FInferType>("FInferType", ElemwiseType<1, 1>)
+.set_attr<nnvm::FGradient>("FGradient", ElemwiseGradUseNone{ "_backward_copy" })
+.set_attr<FCompute>("FCompute<cpu>", IdentityCompute<cpu>)
+.set_attr<nnvm::FInplaceOption>("FInplaceOption",
+  [](const NodeAttrs& attrs) {
+  return std::vector<std::pair<int, int> >{{0, 0}};
+})
+.add_argument("data", "NDArray", "Input data to reshape.");
 
 NNVM_REGISTER_OP(transpose)
 .MXNET_DESCRIBE("Transpose the input tensor and return a new one")
